Update Windows benchmarks with a notation on pending changes.
We're in the process of updating the Linux and Windows benchmarks.  The Linux benchmarks are ready, the Windows ones will be soon.  We want to get the Linux ones out now, so we're adding a note to the Windows benchmarks to let them know of the changes to Linux and pending update to Windows.